Manga_image_scrapper/requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
Pillow==10.2.0 